+++ server/setup.cfg
@@ -1,4 +1,4 @@
 [flake8]
 max-line-length=120
 exclude=env,bin,lib,include,src
-ignore=F811
+ignore=F811,D200,D202,D205,D400,D401,D100,D101,D102,D103,D104,D105,D107,W503,W504,W605,F401,E261,F841
